refactor: move bip32, types, and derivation into wallet/
These files are only used within the wallet module. Moving them
co-locates them with their consumers and simplifies imports.
