+++ src/index.ts
@@ -1,10 +1,12 @@
-export * from './derivation'
 export * from './notes'
 export * from './keys'
 export * from './encoding'
-export type * from './types'
+
+export * from './wallet/derivation'
+export * from './wallet/bip32'
+export type * from './wallet/types'
+export type * from './wallet/wallet-node'
 export { WalletNode } from './wallet/wallet-node'
-export type { WalletNodes } from './wallet/wallet-node'
 export { RailgunWallet } from './wallet/railgun-wallet'
+
 export { Mnemonic } from './mnemonic'
-export { getMasterKeyFromSeed, childKeyDerivationHardened, getPathSegments } from './bip32'
